+++ R/smvr.R
@@ -3,10 +3,20 @@
 #' @description
 #' The `smvr` class represents versions that follow the
 #' [Semantic Versioning Specification (SemVer)](https://semver.org/).
+#' A version number contains three components, `MAJOR.MINOR.PATCH`,
+#' and optional pre-release and build metadata labels.
 #'
-#' - [smvr()] is a constructor for creating `smvr` objects
-#'   from each component.
-#' - [parse_semver()] parses a character vector into `smvr` objects.
+#' This is similar to the base R's [numeric_version] class, but always has
+#' three components (major, minor, patch) and supports pre-release
+#' and build metadata labels. And, unlike [numeric_version],
+#' SemVer uses dots (`.`) as separators and does not allow hyphens (`-`)
+#' except to indicate the start of a pre-release label.
+#'
+#' There are two functions to create [smvr] objects:
+#'
+#' - [smvr()] is a constructor from each component.
+#'   Each component must have the same length or length 1 (will be recycled).
+#' - [parse_semver()] parses a character vector.
 #'
 #' @details
 #' Build metadata is not used for ordering, but the `==` and `!=` operators
@@ -16,26 +26,28 @@
 #'   the major, minor, and patch version components.
 #'   The default values for `minor` and `patch` are `0`.
 #' @param pre_release Something that can be cast to a [pre_release_ids] vector.
-#'   This represents pre-release identifiers, which can be empty (`""`)
-#'   meaning non pre-release.
+#'   This can be empty (`""`) meaning non pre-release (default).
 #' @param build Optional build metadata character vector.
 #'   Should have the pattern `^[a-zA-Z0-9-]+` and can contain
-#'   multiple components separated by dots (`"."`).
-#'   This can be empty (`""`) meaning no build metadata.
+#'   multiple components separated by dots (`.`).
+#'   This can be empty (`""`) meaning no build metadata (default).
 #' @return A [smvr] class vector.
+#' @seealso
+#' - [as_smvr()] to convert other classes to [smvr].
+#' - [extract-component] functions to extract components of a [smvr] object.
+#'   (Operations opposite to [smvr()]).
+#' - [update-version] functions to update components of a [smvr] object.
 #' @examples
 #' # SemVer versions from components
 #' smvr(4, 1:5)
 #'
 #' # Parse SemVer versions from character
-#' parse_semver(c("1.0.0-alpha", "1.0.0-beta+exp.sha.5114f85"))
-#'
 #' v <- parse_semver(c(
 #'   "1.0.0",
 #'   "1.0.0-alpha",
 #'   "1.0.0-beta",
-#'   "1.0.0-rc.1",
 #'   "1.0.0-rc.2",
+#'   "1.0.0-rc.10",
 #'   NA
 #' ))
 #' v
@@ -43,8 +55,8 @@
 #' # Sorting
 #' vctrs::vec_sort(v)
 #'
-#' # Works with base R vectors.
-#' v[v >= "1.0.0-rc.2"]
+#' # Can be compared with string notation
+#' v[v >= "1.0.0-rc.2" & !is.na(v)]
 #'
 #' # Partial version components are treated as NA
 #' suppressWarnings(parse_semver("1.5"))
@@ -53,7 +65,12 @@
 #' # less than 3 components, and can be cast to smvr.
 #' numeric_version("1.5") |>
 #'   vctrs::vec_cast(smvr())
-#' @order 2
+#'
+#' # Be careful with hyphens in numeric_version and SemVer.
+#' # The following examples yield opposite results.
+#' numeric_version("1.0.0-1") > "1.0.0" # 1.0.0-1 is the same as 1.0.0.1
+#' parse_semver("1.0.0-1") > "1.0.0"    # 1.0.0-1 is a pre-release version
+#' @order 1
 #' @export
 smvr <- function(
   major = integer(),
@@ -66,7 +83,7 @@ smvr <- function(
   minor <- vec_cast(minor, integer())
   patch <- vec_cast(patch, integer())
   pre_release <- vec_cast(pre_release, new_pre_release_ids())
-  build <- vec_cast(build, character())
+  build <- vec_cast(build, new_parsed_chr_build_metadata())
 
   if (any(major < 0L, na.rm = TRUE)) {
     cli::cli_abort(
@@ -89,26 +106,6 @@ smvr <- function(
       c(
         "{.code patch} must be non-negative integer values.",
         x = "Problematic values: {.val {patch[patch < 0L]}}"
-      )
-    )
-  }
-  # Check the build metadata pattern
-  if (
-    any(
-      !grepl(BUILD_METADATA_PATTERN, build, perl = TRUE) &
-        nzchar(build) &
-        !is.na(build)
-    )
-  ) {
-    problematic_builds <- build[
-      !grepl(BUILD_METADATA_PATTERN, build, perl = TRUE) &
-        nzchar(build) &
-        !is.na(build)
-    ]
-    cli::cli_abort(
-      c(
-        "{.arg build} must match the pattern {.str {BUILD_METADATA_PATTERN}}.",
-        x = "Problematic values: {.val {problematic_builds}}"
       )
     )
   }
@@ -186,10 +183,16 @@ vec_ptype2.logical.smvr <- function(x, y, ...) smvr()
 #' @export
 vec_ptype2.smvr.logical <- function(x, y, ...) smvr()
 
+# We can't determine the number of identifiers included in the
+# character vector, so only 5 identifiers are allowed.
 #' @export
-vec_ptype2.character.smvr <- function(x, y, ...) smvr()
+vec_ptype2.character.smvr <- function(x, y, ...) {
+  smvr(pre_release = ptype2_chr_ids_impl(x, y))
+}
 #' @export
-vec_ptype2.smvr.character <- function(x, y, ...) smvr()
+vec_ptype2.smvr.character <- function(x, y, ...) {
+  smvr(pre_release = ptype2_chr_ids_impl(y, x))
+}
 
 #' @export
 vec_ptype2.numeric_version.smvr <- function(x, y, ...) {
@@ -220,7 +223,7 @@ vec_cast.smvr.logical <- function(x, to, ...) {
   if (all(is.na(x))) {
     smvr(x)
   } else {
-    cli::cli_abort("Cannot cast non-NA logical to smvr")
+    cli::cli_abort("Cannot cast non-NA logical to {.cls smvr}")
   }
 }
 
@@ -231,7 +234,8 @@ vec_cast.character.smvr <- function(x, to, ...) {
 
 #' @export
 vec_cast.smvr.character <- function(x, to, ...) {
-  parse_semver(x)
+  parse_semver(x) |>
+    vec_cast(to)
 }
 
 #' @export
@@ -239,7 +243,7 @@ vec_cast.numeric_version.smvr <- function(x, to, ...) {
   if (any(is_pre_release(x), na.rm = TRUE)) {
     cli::cli_abort(
       c(
-        "Pre-release {.code smvr} cannot be cast to {.code numeric_version}.",
+        "Pre-release {.cls smvr} cannot be cast to {.cls numeric_version}.",
         x = "Problematic values: {.val {x[is_pre_release(x)]}}"
       )
     )
@@ -258,7 +262,7 @@ vec_cast.smvr.numeric_version <- function(x, to, ...) {
   if (any(n_components > 3L)) {
     cli::cli_abort(
       c(
-        "Cannot cast {.code numeric_version} with more than 3 components to {.code smvr}.",
+        "Cannot cast {.cls numeric_version} with more than 3 components to {.cls smvr}.",
         x = "Problematic values: {.val {x[n_components > 3L]}}"
       )
     )